+++ bc/utils/utils.py
@@ -5,7 +5,7 @@ def is_number(s):
     try:
         float(s)
         return True
-    except (ValueError, TypeError):
+    except (IndexError, ValueError, TypeError):
         return False